@@ -4,74 +4,74 @@
   "scripts": {
     "ng": "ng",
     "start:dev": "ng serve --port 11340",
-    "build": "ng build --configuration production",
+    "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.2",
-    "@angular/cdk": "~13.3.2",
-    "@angular/common": "~13.3.2",
-    "@angular/compiler": "~13.3.2",
-    "@angular/core": "~13.3.2",
-    "@angular/forms": "~13.3.2",
-    "@angular/localize": "~13.3.2",
-    "@angular/platform-browser": "~13.3.2",
-    "@angular/platform-browser-dynamic": "~13.3.2",
-    "@angular/router": "~13.3.2",
-    "@angular/service-worker": "~13.3.2",
+    "@angular/animations": "^14.0.3",
+    "@angular/cdk": "^14.0.3",
+    "@angular/common": "^14.0.3",
+    "@angular/compiler": "^14.0.3",
+    "@angular/core": "^14.0.3",
+    "@angular/forms": "^14.0.3",
+    "@angular/localize": "^14.0.3",
+    "@angular/platform-browser": "^14.0.3",
+    "@angular/platform-browser-dynamic": "^14.0.3",
+    "@angular/router": "^14.0.3",
+    "@angular/service-worker": "^14.0.3",
     "@ctrl/ngx-codemirror": "~5.1.1",
-    "@ng-bootstrap/ng-bootstrap": "^12.0.2",
+    "@ng-bootstrap/ng-bootstrap": "^12.1.2",
     "@popperjs/core": "^2.11.5",
-    "angular-gridster2": "^13.2.0",
+    "angular-gridster2": "^14.0.1",
     "angular-split": "^13.2.0",
     "bootstrap": "~5.1.3",
-    "bootstrap-darkmode": "^5.0.0-beta.1",
-    "bootstrap-icons": "^1.8.1",
+    "bootstrap-darkmode": "^5.0.1",
+    "bootstrap-icons": "^1.8.3",
     "bson-objectid": "^2.0.3",
-    "codemirror": "^5.65.2",
+    "codemirror": "^5.65.6",
     "file-saver": "^2.0.5",
-    "highlight.js": "^11.5.0",
-    "keycloak-angular": "^9.1.0",
-    "keycloak-js": "15.1.0",
-    "ng-bootstrap-darkmode": "^0.5.0",
-    "ng-bootstrap-ext": "^0.3.0-beta.2",
-    "ng-terminal": "^5.0.0-beta.2",
+    "highlight.js": "^11.5.1",
+    "keycloak-angular": "^10.0.1",
+    "keycloak-js": "18.0.1",
+    "ng-bootstrap-darkmode": "^5.0.0",
+    "ng-bootstrap-ext": "^0.4.1",
+    "ng-terminal": "^5.0.0",
     "ngx-clipboard": "^15.1.0",
     "ngx-drag-drop": "^13.1.0-alpha.1",
     "rxjs": "~7.5.5",
-    "tslib": "^2.3.1",
-    "zone.js": "~0.11.5"
+    "tslib": "^2.4.0",
+    "zone.js": "~0.11.6"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.2",
-    "@angular-eslint/builder": "~13.2.0",
-    "@angular-eslint/eslint-plugin": "~13.2.0",
-    "@angular-eslint/eslint-plugin-template": "~13.2.0",
-    "@angular-eslint/schematics": "~13.2.0",
-    "@angular-eslint/template-parser": "~13.2.0",
-    "@angular/cli": "~13.3.2",
-    "@angular/compiler-cli": "~13.3.2",
-    "@angular/language-service": "~13.3.2",
+    "@angular-devkit/build-angular": "^14.0.3",
+    "@angular-eslint/builder": "~14.0.0",
+    "@angular-eslint/eslint-plugin": "~14.0.0",
+    "@angular-eslint/eslint-plugin-template": "~14.0.0",
+    "@angular-eslint/schematics": "~14.0.0",
+    "@angular-eslint/template-parser": "~14.0.0",
+    "@angular/cli": "^14.0.3",
+    "@angular/compiler-cli": "^14.0.3",
+    "@angular/language-service": "^14.0.3",
     "@types/codemirror": "^5.60.5",
     "@types/file-saver": "^2.0.5",
     "@types/jasmine": "~3.6.11",
     "@types/jasminewd2": "~2.0.10",
-    "@types/node": "^16.11.26",
-    "@typescript-eslint/eslint-plugin": "^5.18.0",
-    "@typescript-eslint/parser": "^5.18.0",
-    "eslint": "^8.12.0",
-    "jasmine-core": "~3.6.0",
+    "@types/node": "^16.11.41",
+    "@typescript-eslint/eslint-plugin": "^5.29.0",
+    "@typescript-eslint/parser": "^5.29.0",
+    "eslint": "^8.18.0",
+    "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~5.0.2",
-    "karma": "~6.3.17",
+    "karma": "~6.3.20",
     "karma-chrome-launcher": "~3.1.1",
     "karma-coverage-istanbul-reporter": "~3.0.3",
     "karma-jasmine": "~4.0.2",
     "karma-jasmine-html-reporter": "^1.7.0",
     "protractor": "~7.0.0",
-    "ts-node": "^10.7.0",
-    "typescript": "~4.6.3"
+    "ts-node": "^10.8.1",
+    "typescript": "~4.6.4"
   }
 }
