@@ -21,62 +21,64 @@
   },
   "dependencies": {
     "@elastic/elasticsearch": "~7.13.0",
-    "@nestjs/axios": "^0.0.7",
-    "@nestjs/common": "^8.4.3",
-    "@nestjs/core": "^8.4.3",
+    "@nestjs/axios": "^0.0.8",
+    "@nestjs/common": "^8.4.7",
+    "@nestjs/core": "^8.4.7",
     "@nestjs/elasticsearch": "^8.1.0",
-    "@nestjs/event-emitter": "^1.1.0",
-    "@nestjs/jwt": "^8.0.0",
-    "@nestjs/microservices": "^8.4.3",
-    "@nestjs/mongoose": "^9.0.3",
-    "@nestjs/passport": "^8.2.1",
-    "@nestjs/platform-express": "^8.4.3",
-    "@nestjs/schedule": "^1.1.0",
+    "@nestjs/event-emitter": "^1.1.1",
+    "@nestjs/jwt": "^8.0.1",
+    "@nestjs/microservices": "^8.4.7",
+    "@nestjs/mongoose": "^9.1.1",
+    "@nestjs/passport": "^8.2.2",
+    "@nestjs/platform-express": "^8.4.7",
+    "@nestjs/schedule": "^2.0.1",
     "@nestjs/swagger": "^5.2.1",
-    "@types/dockerode": "^3.3.8",
+    "@types/dockerode": "^3.3.9",
     "@types/passport-jwt": "^3.0.6",
-    "axios": "^0.26.1",
+    "axios": "^0.27.2",
     "bson-objectid": "^2.0.3",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.13.2",
-    "dockerode": "^3.3.1",
+    "dockerode": "^3.3.2",
+    "express": "^4.18.1",
     "glob-to-regexp": "^0.4.1",
-    "mongoose": "^6.2.10",
-    "multer": "^1.4.4",
-    "nats": "^2.6.1",
-    "passport": "^0.5.2",
+    "mongoose": "^6.4.0",
+    "multer": "^1.4.5-lts.1",
+    "nats": "^2.7.1",
+    "passport": "^0.5.3",
     "passport-jwt": "^4.0.0",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "rxjs": "^7.5.5",
-    "swagger-ui-express": "^4.3.0",
+    "swagger-ui-express": "^4.4.0",
     "unzipper": "^0.10.11"
   },
   "devDependencies": {
-    "@nestjs/cli": "^8.2.5",
-    "@nestjs/schematics": "^8.0.10",
-    "@nestjs/testing": "^8.4.3",
-    "@types/cron": "^1.7.3",
+    "@nestjs/cli": "^8.2.8",
+    "@nestjs/schematics": "^8.0.11",
+    "@nestjs/testing": "^8.4.7",
+    "@types/cron": "^2.0.0",
     "@types/express": "^4.17.13",
     "@types/glob-to-regexp": "^0.4.1",
-    "@types/jest": "^27.4.1",
+    "@types/jest": "^27.5.2",
     "@types/multer": "^1.4.7",
-    "@types/node": "^16.11.26",
+    "@types/node": "^16.11.41",
     "@types/supertest": "^2.0.12",
     "@types/unzipper": "^0.10.5",
-    "@typescript-eslint/eslint-plugin": "^5.18.0",
-    "@typescript-eslint/parser": "^5.18.0",
-    "eslint": "^8.12.0",
+    "@typescript-eslint/eslint-plugin": "^5.29.0",
+    "@typescript-eslint/parser": "^5.29.0",
+    "eslint": "^8.18.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
     "jest": "^27.5.1",
-    "prettier": "^2.6.2",
-    "supertest": "^6.2.2",
-    "ts-jest": "^27.1.4",
-    "ts-loader": "^9.2.8",
-    "ts-node": "^10.7.0",
+    "prettier": "^2.7.1",
+    "supertest": "^6.2.3",
+    "ts-jest": "^27.1.5",
+    "ts-loader": "^9.3.1",
+    "ts-node": "^10.8.1",
     "tsconfig-paths": "^3.14.1",
-    "typescript": "^4.6.3"
+    "typescript": "^4.7.4",
+    "webpack": "^5.73.0"
   },
   "jest": {
     "moduleFileExtensions": [
