@@ -1,11 +1,14 @@
 import {Component, OnDestroy, OnInit, TemplateRef, ViewChild} from '@angular/core';
 import {ActivatedRoute, NavigationEnd, Router} from '@angular/router';
 import {KeycloakService} from 'keycloak-angular';
-import {Subscription} from 'rxjs';
-import {filter, startWith, switchMap} from 'rxjs/operators';
+import { ToastService } from 'ng-bootstrap-ext';
+import {forkJoin, of, Subscription} from 'rxjs';
+import {catchError, filter, startWith, switchMap, tap} from 'rxjs/operators';
 import {User} from '../../../user/user';
 import {UserService} from '../../../user/user.service';
+import {Container} from '../../model/container';
 import {LocalProject, Project} from '../../model/project';
+import { ContainerService } from '../../services/container.service';
 import {MemberService} from '../../services/member.service';
 import {ProjectService} from '../../services/project.service';
 
@@ -17,9 +20,10 @@ import {ProjectService} from '../../services/project.service';
 export class ProjectListComponent implements OnInit, OnDestroy {
   @ViewChild('editModal', {static: true}) editModal: TemplateRef<any>;
 
-  currentUser?: User;
+  currentUser?: User | null;
 
   projects: Project[] = [];
+  containers: (Container | null)[] = [];
 
   private subscription: Subscription;
 
@@ -30,19 +34,27 @@ export class ProjectListComponent implements OnInit, OnDestroy {
     private userService: UserService,
     private memberService: MemberService,
     private keycloak: KeycloakService,
+    private containerService: ContainerService,
+    private toastService: ToastService,
   ) {
   }
 
   ngOnInit(): void {
-    this.userService.current$.subscribe(user => user && (this.currentUser = user));
 
     this.subscription = this.router.events.pipe(
       filter(e => e instanceof NavigationEnd && e.urlAfterRedirects === '/projects'),
       startWith(undefined),
       switchMap(() => this.projectService.getOwn()),
-    ).subscribe(projects => {
-      this.projects = projects;
-    });
+      tap(projects => this.projects = projects),
+      switchMap(projects => forkJoin(projects.map(project => this.containerService.get(project.id).pipe(
+        catchError(() => of(null)),
+      )))),
+      tap(containers => this.containers = containers),
+    ).subscribe();
+
+    this.subscription.add(
+      this.userService.current$.subscribe(user => this.currentUser = user),
+    );
   }
 
   ngOnDestroy() {
@@ -59,6 +71,9 @@ export class ProjectListComponent implements OnInit, OnDestroy {
       if (index >= 0) {
         this.projects[index] = persistentProject;
       }
+      this.toastService.success('Convert Project', 'Successfully converted project to persistent.');
+    }, error => {
+      this.toastService.error('Convert Project', 'Failed to convert project', error);
     });
   }
 
@@ -69,6 +84,9 @@ export class ProjectListComponent implements OnInit, OnDestroy {
 
     this.memberService.delete({projectId: project.id, userId: this.currentUser!.id!}).subscribe(() => {
       this.projects.removeFirst(p => p === project);
+      this.toastService.warn('Leave Project', 'Successfully left project.');
+    }, error => {
+      this.toastService.error('Leave Project', 'Failed to leave project', error);
     });
   }
 }
