@@ -1,12 +1,12 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
 
-import {AboutComponent} from './about/about.component';
-import {ChangelogComponent} from './changelog/changelog.component';
-import {FeedbackComponent} from './feedback/feedback.component';
-import {HomeComponent} from './home/home.component';
-import {PageNotFoundComponent} from './page-not-found/page-not-found.component';
-import {PrivacyComponent} from './privacy/privacy.component';
+import {AboutComponent} from './components/about/about.component';
+import {ChangelogComponent} from './components/changelog/changelog.component';
+import {FeedbackComponent} from './components/feedback/feedback.component';
+import {HomeComponent} from './components/home/home.component';
+import {PageNotFoundComponent} from './components/page-not-found/page-not-found.component';
+import {PrivacyComponent} from './components/privacy/privacy.component';
 
 const routes: Routes = [
   {path: 'scenarios', loadChildren: () => import('./editor/editor.module').then(m => m.EditorModule)},
