@@ -8,21 +8,21 @@ import {ServiceWorkerModule} from '@angular/service-worker';
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
 import {KeycloakAngularModule, KeycloakService} from 'keycloak-angular';
 import {NgBootstrapDarkmodeModule, THEME_LOADER, THEME_SAVER, ThemeLoader, ThemeSaver} from 'ng-bootstrap-darkmode';
-import {ToastModule} from 'ng-bootstrap-ext';
+import {ModalModule, ToastModule} from 'ng-bootstrap-ext';
 import {of} from 'rxjs';
 
 import {environment} from '../environments/environment';
 
-import {AboutComponent} from './about/about.component';
+import {AboutComponent} from './components/about/about.component';
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
-import {ChangelogComponent} from './changelog/changelog.component';
-import {FeedbackComponent} from './feedback/feedback.component';
-import {HeaderComponent} from './header/header.component';
-import {HomeComponent} from './home/home.component';
-import {PageNotFoundComponent} from './page-not-found/page-not-found.component';
-import {PrivacyService} from './privacy.service';
-import {PrivacyComponent} from './privacy/privacy.component';
+import {ChangelogComponent} from './components/changelog/changelog.component';
+import {FeedbackComponent} from './components/feedback/feedback.component';
+import {HeaderComponent} from './components/header/header.component';
+import {HomeComponent} from './components/home/home.component';
+import {PageNotFoundComponent} from './components/page-not-found/page-not-found.component';
+import {PrivacyService} from './services/privacy.service';
+import {PrivacyComponent} from './components/privacy/privacy.component';
 import {SharedModule} from './shared/shared.module';
 import {UserModule} from './user/user.module';
 
@@ -74,6 +74,7 @@ function initializeKeycloak(keycloak: KeycloakService) {
       // or after 30 seconds (whichever comes first).
       registrationStrategy: 'registerWhenStable:30000',
     }),
+    ModalModule,
   ],
   providers: [
     {
