@@ -3,7 +3,7 @@ import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
 
 import {environment} from '../../environments/environment';
-import {PrivacyService} from '../privacy.service';
+import {PrivacyService} from '../services/privacy.service';
 import {Panel} from './model/panel';
 import {Request} from './model/request';
 import {Response} from './model/response';
