@@ -2,5 +2,5 @@ export interface GenerateResult {
   board: string;
   pages?: string[];
   diagrams?: string[];
-  classDiagram?: string;
+  classDiagrams?: string[];
 }
