@@ -1,13 +1,15 @@
 import {Prop, Schema, SchemaFactory} from '@nestjs/mongoose';
-import {ApiProperty} from '@nestjs/swagger';
-import {IsArray, IsMongoId, IsNotEmpty, IsString} from 'class-validator';
+import {ApiProperty, ApiPropertyOptional} from '@nestjs/swagger';
+import {IsArray, IsMongoId, IsNotEmpty, IsOptional, IsString, IsUUID} from 'class-validator';
 import {Document} from 'mongoose';
 
 @Schema()
 export class Course {
   @Prop()
-  @ApiProperty()
-  createdBy: string;
+  @ApiPropertyOptional()
+  @IsOptional()
+  @IsUUID()
+  createdBy?: string;
 
   @Prop()
   @ApiProperty()
