@@ -13,7 +13,7 @@ export const environment = {
     version: process.env.DOCKER_VERSION || 'v1.41',
     bindPrefix: process.env.FULIB_PROJECTS_DATA_DIR || 'data',
     containerImage: process.env.FULIB_PROJECTS_CONTAINER_IMAGE || 'codercom/code-server:latest',
-    proxyHost: process.env.FULIB_PROJECTS_PROXY_URL || 'http://localhost:8080',
+    proxyHost: process.env.FULIB_PROJECTS_PROXY_URL || 'http://localhost:13147',
     network: process.env.FULIB_PROJECTS_NETWORK || 'fuliborg_projects',
   },
   auth: {
